+++ mobile/android/android-components/components/browser/session/src/main/java/mozilla/components/browser/session/ext/SessionExtensions.kt
@@ -47,8 +47,7 @@ private fun Session.toContentState(): ContentState {
         title,
         progress,
         loading,
-        searchTerms,
-        securityInfo.toSecurityInfoState()
+        securityInfo = securityInfo.toSecurityInfoState()
     )
 }