+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/TabsUseCases.kt
@@ -136,6 +136,7 @@ class TabsUseCases(
          * @param flags the [LoadUrlFlags] to use when loading the provided URL.
          * @param contextId the session context id to use for this tab.
          * @param engineSession (optional) engine session to use for this tab.
+         * @return The ID of the created tab.
          */
         @Suppress("LongParameterList")
         operator fun invoke(
@@ -145,9 +146,11 @@ class TabsUseCases(
             parentId: String? = null,
             flags: LoadUrlFlags = LoadUrlFlags.none(),
             contextId: String? = null,
-            engineSession: EngineSession? = null
-        ): Session {
-            val session = Session(url, false, Source.NEW_TAB, contextId = contextId)
+            engineSession: EngineSession? = null,
+            source: Source = Source.NEW_TAB
+        ): String {
+            val session = Session(url, false, source, contextId = contextId)
+
             val parent = parentId?.let { sessionManager.findSessionById(parentId) }
             sessionManager.add(session, selected = selectTab, engineSession = engineSession, parent = parent)
 
@@ -161,7 +164,7 @@ class TabsUseCases(
                 ))
             }
 
-            return session
+            return session.id
         }
     }
 
@@ -189,6 +192,7 @@ class TabsUseCases(
          * @param parentId the id of the parent tab to use for the newly created tab.
          * @param flags the [LoadUrlFlags] to use when loading the provided URL.
          * @param engineSession (optional) engine session to use for this tab.
+         * @return The ID of the created tab.
          */
         @Suppress("LongParameterList")
         operator fun invoke(
@@ -197,9 +201,11 @@ class TabsUseCases(
             startLoading: Boolean = true,
             parentId: String? = null,
             flags: LoadUrlFlags = LoadUrlFlags.none(),
-            engineSession: EngineSession? = null
-        ): Session {
-            val session = Session(url, true, Source.NEW_TAB)
+            engineSession: EngineSession? = null,
+            source: Source = Source.NEW_TAB
+        ): String {
+            val session = Session(url, true, source)
+
             val parent = parentId?.let { sessionManager.findSessionById(parentId) }
             sessionManager.add(session, selected = selectTab, engineSession = engineSession, parent = parent)
 
@@ -213,7 +219,7 @@ class TabsUseCases(
                 ))
             }
 
-            return session
+            return session.id
         }
     }